Deprecated internal methods for UUID type
